+++ apps/example/src/api/person.ts
@@ -1,36 +1,12 @@
 import request from "@Passets/request";
 import type { PageResult } from "@Passets/request";
+import type {
+  PersonItem,
+  PersonListParams,
+  PersonPayload,
+} from "./person.types";
 
-/** 示例人员记录 */
-export interface PersonItem {
-  id: number;
-  name: string;
-  age: number;
-  sex: string;
-  ethnic: string;
-  isHealthy: string;
-}
-
-/** 人员列表查询参数：分页必填，其余为可选搜索条件 */
-export interface PersonListParams {
-  pageNumber: number;
-  pageSize: number;
-  name?: string;
-  age?: number;
-  sex?: string;
-  ethnic?: string;
-  isHealthy?: string;
-}
-
-/** 人员新增/编辑入参 */
-export interface PersonPayload {
-  id?: number;
-  name: string;
-  age: number;
-  sex: string;
-  ethnic: string;
-  isHealthy: string;
-}
+export * from "./person.types";
 
 export const getPersonList = (data: PersonListParams) =>
   request.post<PageResult<PersonItem>>({ url: "/example/person/getList", data });
